requirements.txt
discord.py>=2.0.0
aiohttp
beautifulsoup4
python-dotenv

